refactor(ir): split intrinsic LIR instructions
